refactor functions to util.js
